+++ doc/html/search/defines_8.js
@@ -1,12 +1,5 @@
 var searchData=
 [
   ['halomod_5ftolerance',['HALOMOD_TOLERANCE',['../ccl__test__halomod_8c.html#a13d7f1fae1dedeed78a89c0a1d115e3d',1,'ccl_test_halomod.c']]],
-  ['hh',['HH',['../ccl__3dcorr__sample_8c.html#a98e0283d49017dc771669a2a07dbdcf8',1,'HH():&#160;ccl_3dcorr_sample.c'],['../ccl__corr__sample_8c.html#a98e0283d49017dc771669a2a07dbdcf8',1,'HH():&#160;ccl_corr_sample.c'],['../ccl__sample___omnu_8c.html#a98e0283d49017dc771669a2a07dbdcf8',1,'HH():&#160;ccl_sample_Omnu.c'],['../ccl__test__correlation__3dbis_8c.html#a98e0283d49017dc771669a2a07dbdcf8',1,'HH():&#160;ccl_test_correlation_3dbis.c'],['../pxi__fidtest_8c.html#a98e0283d49017dc771669a2a07dbdcf8',1,'HH():&#160;pxi_fidtest.c'],['../pxi__test_8c.html#a98e0283d49017dc771669a2a07dbdcf8',1,'HH():&#160;pxi_test.c']]],
-  ['hm_5fepsabs',['HM_EPSABS',['../ccl__halomod_8h.html#aa45e5826c69d5f458325cbab3e132e4e',1,'ccl_halomod.h']]],
-  ['hm_5fepsrel',['HM_EPSREL',['../ccl__halomod_8h.html#a42d823f95333d2af4af5b7cdb99224ff',1,'ccl_halomod.h']]],
-  ['hm_5fint_5fmethod',['HM_INT_METHOD',['../ccl__halomod_8h.html#af180b285f420f4ef91cbfd8e2ba994cf',1,'ccl_halomod.h']]],
-  ['hm_5flimit',['HM_LIMIT',['../ccl__halomod_8h.html#a6f707fcb7566b2288488b671a9b6bf04',1,'ccl_halomod.h']]],
-  ['hm_5fmmax',['HM_MMAX',['../ccl__halomod_8h.html#a2a4961f32cf3f0ebbc255fbb4018ae7a',1,'ccl_halomod.h']]],
-  ['hm_5fmmin',['HM_MMIN',['../ccl__halomod_8h.html#a43280101e6b08728406a683537e55539',1,'ccl_halomod.h']]],
-  ['hplanck',['HPLANCK',['../ccl__constants_8h.html#a32c64516e5ad4b3ec65b54deb3f27457',1,'ccl_constants.h']]]
+  ['hh',['HH',['../ccl__3dcorr__sample_8c.html#a98e0283d49017dc771669a2a07dbdcf8',1,'HH():&#160;ccl_3dcorr_sample.c'],['../ccl__corr__sample_8c.html#a98e0283d49017dc771669a2a07dbdcf8',1,'HH():&#160;ccl_corr_sample.c']]]
 ];