+++ doc/html/search/enums_1.js
@@ -2,6 +2,8 @@ var searchData=
 [
   ['ccl_5fmnu_5fconvention',['ccl_mnu_convention',['../ccl__core_8h.html#a8c23ee4dd2faea6685776fc39c32104b',1,'ccl_core.h']]],
   ['ccl_5fneutrino_5fmass_5fsplits',['ccl_neutrino_mass_splits',['../ccl__neutrinos_8h.html#a123045c726e0b48e6cc308e48554ec44',1,'ccl_neutrinos.h']]],
+  ['ccl_5fp2d_5fextrap_5fgrowth_5ft',['ccl_p2d_extrap_growth_t',['../ccl__core_8h.html#a7c3bb0ab563e8da9e642173d5abdb208',1,'ccl_core.h']]],
+  ['ccl_5fp2d_5finterp_5ft',['ccl_p2d_interp_t',['../ccl__core_8h.html#ae3cd178ed38927c026efe52f72cad865',1,'ccl_core.h']]],
   ['ccl_5fspecies_5fx_5flabel',['ccl_species_x_label',['../ccl__background_8h.html#a3ab9842523d3f98e1ab5c57f0249a15d',1,'ccl_background.h']]],
   ['ccl_5ftracer_5ffunc_5ft',['ccl_tracer_func_t',['../ccl__cls_8h.html#ad40f84d4342c4ba1b55267186a8002e8',1,'ccl_cls.h']]],
   ['ccl_5ftracer_5ft',['ccl_tracer_t',['../ccl__cls_8h.html#ad89f3ee4117e3c2eeff11256df700ea6',1,'ccl_cls.h']]],