+++ doc/html/search/variables_14.js
@@ -1,9 +1,5 @@
 var searchData=
 [
   ['y0',['y0',['../struct_spl_par.html#a49d3136602dec3eff0bfd5c631615cb0',1,'SplPar']]],
-  ['yf',['yf',['../struct_spl_par.html#a3e483146b3649beeaa42e854f7b7a386',1,'SplPar']]],
-  ['your_5fdn_5ffunc',['your_dN_func',['../structd_ndz__info.html#a2e30312218187d41c9794ebeb51a929e',1,'dNdz_info']]],
-  ['your_5fdn_5fparams',['your_dN_params',['../structd_ndz__info.html#aee9a21a14c1bfcd3243eb9d1ab462ac8',1,'dNdz_info']]],
-  ['your_5fpz_5ffunc',['your_pz_func',['../structpz__info.html#a9ffc35097b7d9bd3dee298991f750a1a',1,'pz_info']]],
-  ['your_5fpz_5fparams',['your_pz_params',['../structpz__info.html#a8ca7c69c664628da1d8ac4ed445293a7',1,'pz_info']]]
+  ['yf',['yf',['../struct_spl_par.html#a3e483146b3649beeaa42e854f7b7a386',1,'SplPar']]]
 ];