+++ tests/ccl_test_pyccl_interface.py
@@ -440,73 +440,6 @@ def check_neutrinos():
                                              m_nu=42)
 
 
-def check_redshifts(cosmo):
-    """
-    Check that redshift functions can be run and produce finite values.
-    """
-    # Types of scale factor input (scalar, list, array)
-    a_scl = 0.5
-    a_lst = [0.2, 0.4, 0.6, 0.8, 1.]
-    a_arr = np.linspace(0.2, 1., 5)
-
-    # Types of redshift input
-    z_scl = 0.5
-    z_lst = [0., 0.5, 1., 1.5, 2.]
-    z_arr = np.array(z_lst)
-
-    # p(z) function for dNdz_tomog
-    def pz1(z_ph, z_s, args):
-        return np.exp(- (z_ph - z_s)**2. / 2.)
-
-    # Lambda function p(z) function for dNdz_tomog
-    pz2 = lambda z_ph, z_s, args: np.exp(-(z_ph - z_s)**2. / 2.)
-
-    # PhotoZFunction classes
-    PZ1 = ccl.PhotoZFunction(pz1)
-    PZ2 = ccl.PhotoZFunction(pz2)
-    PZ3 = ccl.PhotoZGaussian(sigma_z0=0.1)
-
-    # dNdz (in terms of true redshift) function for dNdz_tomog
-    def dndz1(z, args):
-        return z**1.24 * np.exp(- (z / 0.51)**1.01)
-    # dNdzFunction classes
-    dNdZ1 = ccl.dNdzFunction(dndz1)
-    dNdZ2 = ccl.dNdzSmail(alpha = 1.24, beta = 1.01, z0 = 0.51)
-
-    # Check that dNdz_tomog is finite with the various combinations
-    # of PhotoZ and dNdz functions
-    zmin = 0.
-    zmax = 1.
-
-    assert_( all_finite(ccl.dNdz_tomog(z_scl, zmin, zmax, PZ1, dNdZ1)) )
-    assert_( all_finite(ccl.dNdz_tomog(z_lst, zmin, zmax, PZ1, dNdZ1)) )
-    assert_( all_finite(ccl.dNdz_tomog(z_arr, zmin, zmax, PZ1, dNdZ1)) )
-
-    assert_( all_finite(ccl.dNdz_tomog(z_scl, zmin, zmax, PZ2, dNdZ1)) )
-    assert_( all_finite(ccl.dNdz_tomog(z_lst, zmin, zmax, PZ2, dNdZ1)) )
-    assert_( all_finite(ccl.dNdz_tomog(z_arr, zmin, zmax, PZ2, dNdZ1)) )
-
-    assert_( all_finite(ccl.dNdz_tomog(z_scl, zmin, zmax, PZ3, dNdZ1)) )
-    assert_( all_finite(ccl.dNdz_tomog(z_lst, zmin, zmax, PZ3, dNdZ1)) )
-    assert_( all_finite(ccl.dNdz_tomog(z_arr, zmin, zmax, PZ3, dNdZ1)) )
-
-    assert_( all_finite(ccl.dNdz_tomog(z_scl, zmin, zmax, PZ1, dNdZ2)) )
-    assert_( all_finite(ccl.dNdz_tomog(z_lst, zmin, zmax, PZ1, dNdZ2)) )
-    assert_( all_finite(ccl.dNdz_tomog(z_arr, zmin, zmax, PZ1, dNdZ2)) )
-
-    assert_( all_finite(ccl.dNdz_tomog(z_scl, zmin, zmax, PZ2, dNdZ2)) )
-    assert_( all_finite(ccl.dNdz_tomog(z_lst, zmin, zmax, PZ2, dNdZ2)) )
-    assert_( all_finite(ccl.dNdz_tomog(z_arr, zmin, zmax, PZ2, dNdZ2)) )
-
-    assert_( all_finite(ccl.dNdz_tomog(z_scl, zmin, zmax, PZ3, dNdZ2)) )
-    assert_( all_finite(ccl.dNdz_tomog(z_lst, zmin, zmax, PZ3, dNdZ2)) )
-    assert_( all_finite(ccl.dNdz_tomog(z_arr, zmin, zmax, PZ3, dNdZ2)) )
-
-    # Wrong function type
-    assert_raises(TypeError, ccl.dNdz_tomog, z_scl, zmin, zmax, pz1, z_arr)
-    assert_raises(TypeError, ccl.dNdz_tomog, z_scl,  zmin, zmax, z_arr, dNdZ1)
-    assert_raises(TypeError, ccl.dNdz_tomog, z_scl, zmin, zmax, None, None)
-
 def check_cls(cosmo):
     """
     Check that cls functions can be run.
@@ -814,16 +747,6 @@ def test_neutrinos():
     Test neutrino-related functions.
     """
     yield check_neutrinos
-
-def test_redshifts():
-    """
-    Test redshifts module.
-    """
-    for cosmo in reference_models():
-        yield check_redshifts, cosmo
-
-    for cosmo_nu in reference_models_nu():
-       yield check_redshifts, cosmo_nu
 
 @decorators.slow
 def test_cls():