+++ CHANGELOG.md
@@ -5,12 +5,14 @@
 - Fixed a bug where CLASS would crash due to small rounding errors at z = 0
   when evaluating power spectra (#563, #564).
 - Fixed bug in fftlog for some complex arguments (#565, #566).
+- Deprecated the `ccl_redshifts.h` functions (#579).
 - Refactored spline and numerical parameters to be allocated per cosmology (#557).
 - Allow global physical constants to be changed (#557).
 - Fixed memory leaks in `ccl_correlation.c` (#581).
 
 ## Python library
 - Improved error reporting for `angular_cl` computations (#567).
+- Deprecated the `pyccl.redshifts` module (#579).
 - Remove global splines for RSD correlation functions. These are now stored
   per cosmology. Further, they are now rebuilt on-the-fly for a given
   cosmology if a new scale factor is requested. (#582)